fix: removed 304 from http PUT requests --- before this commit, all routes which would update either houses, tasks or contractors; would all respond with 304 (no change) if nothing on the server was updated with the request.
since this code, is outside of 'ok' range (200 -> 299), it would error
out in application. this commit, removes the check all-together.

now all these functions respond with 200, wether or not anything on the
db was changed. (the query is successful to the db, but db deemed that
no change was needed.)