feat/fix: indentity middleware
The identity middleware, which has been lying dormant in `middleware.js`
since it's been written, has finally been completed. The middeleware is
used to identify a user (on index route) if the user is "logged" in
or not.

Users with different roles, have their tokens signed with different
secrets, therefore, the regular `mw.auth()` doesn't work on "index" as
the `mw.auth()` middleware can't really be adapted to fit this change.
(be adapted to check different secrets)

Thusly, if a token is stored in a cookie, it's decoded (no verification)
and the user's "role" is set accordingly (`req.user` is populated).
However the "token" is flagged as "unsafe" and checks have been placed
to ensure that the token is properly validated when navigating further.