feat: modal for removing appointees from task
A modal for removing contractors from a task has been added, utilizing
some CSS from W3Schools for creating a "slider toggle", as I want the
user to pick and choose SEVERAL appointees to remove in a single request.

So far, there is no endpoint for this, but it will be added in a future
commit.