fix: task/suggestion forms <select> not working
In the Worker-View, a contractor could open the suggestion form, and
have the HouseID `<select>` element showing the "selected" house (by
navigating with "Open Form" while viewing the house).

This wouldn't actually be selected, and Alpine would retreive it's id
as null. In the User-View, the same thing could happen.

This has been solved by:
- WorkerView: "retreiving the id properly" before `createSuggestion()`
- UserView: not showing a selected houseID at all