fix: creating a task can't "redir" properly
When creating a task, the "subtab" will be changed to taskDetails,
showing the details for the newly created task.
However, this doesn't work for created tasks, as a newly created task
cannot be selected yet. By re-factoring the way that data is retreived
during task create/update, this has been fixed.