fix: added handling for missing #profileID
The select with id `#profileID` is the dropdown for which a contractor
can choose it's "profile" - hence, this is only present if the user is
a contractor. This results in an error being console.logged in all
other views.

the change - add `?`:

`document.querySelector('#profileID').addEventlist[...]` ->
`document.querySelector('#profileID')?.addEventlis[...]`